Revamped "How To Play" and added Goon Gallery
Updated the "How To Play" function to provide slightly more details, such as how the resulting score is calculated for each game. Added details related to KOs to this section as well.

Added a "What Are Goons?" button to let players see the goon gallery, which has detailed summaries for each type of goon.