Slight refactoring
Removed a function that was largely obsolete.

Moved the vibe display update function to main's _process so that the vibe display is updated as often as possible. I don't quite like this solution, but I'm not sure I can do any better until global signals are implemented. At the very least, the vibes counter is less likely to be incorrect during scoring.